Implement Rand for PieceType
Instead of having a loop that gets a random u8 and continues if a value
outside of [0,6] appears (which, if I have the math right, has an
expected number of iterations of 36), this uses gen_range() to get a
random number in the right range.

Also, by implementing Rand for PieceType, a random PieceType can be
constructed by just random().

Signed-off-by: Uli Schlachter <psychon@znc.in>